Rejigger bindings so that we can wrap for convenience.
The C API isn't natural for Dylan usage in all cases and we need
extra code to cope with that.  We move the generated bindings
from z3-api into %z3-api. We provide some wrappers in z3-api which
also re-exports %z3-api. This allows melange to generate the module
definition for what it is generating.

We then provide wrappers for the functions that take an array of
Z3_ast objects to make using them from Dylan easy.